backend/requirements.txt
Django==3.2.5
django-allauth
django-cors-headers==3.8.0
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
